Directory may not exist
The removed line assumes a `/home` directory exists for the user running pgbouncer, but that's a poor assumption given a lot of system users have their home directory elsewhere.  Simply leaving the WorkingDirectory line out of the systemd configuration causes it to use the given user's home directory by default, sidestepping the need to set it altogether.  From the manpage: "If not set, defaults to the root directory when systemd is running as a system instance and the respective user's home directory if run as user."